+++ sdks/java/io/iceberg/src/main/java/org/apache/beam/sdk/io/iceberg/IcebergCatalogConfig.java
@@ -84,17 +84,26 @@ public abstract class IcebergCatalogConfig implements Serializable {
   }
 
   public void createTable(
-      String tableIdentifier, Schema tableSchema, @Nullable List<String> partitionFields) {
+      String tableIdentifier,
+      Schema tableSchema,
+      @Nullable List<String> partitionFields,
+      @Nullable Map<String, String> tableProperties) {
     TableIdentifier icebergIdentifier = TableIdentifier.parse(tableIdentifier);
     org.apache.iceberg.Schema icebergSchema = IcebergUtils.beamSchemaToIcebergSchema(tableSchema);
     PartitionSpec icebergSpec = PartitionUtils.toPartitionSpec(partitionFields, tableSchema);
     try {
-      catalog().createTable(icebergIdentifier, icebergSchema, icebergSpec);
+      catalog()
+          .createTable(
+              icebergIdentifier,
+              icebergSchema,
+              icebergSpec,
+              tableProperties == null ? Maps.newHashMap() : tableProperties);
       LOG.info(
-          "Created table '{}' with schema: {}\n, partition spec: {}",
+          "Created table '{}' with schema: {}\n, partition spec: {}, table properties: {}",
           icebergIdentifier,
           icebergSchema,
-          icebergSpec);
+          icebergSpec,
+          tableProperties);
     } catch (AlreadyExistsException e) {
       throw new TableAlreadyExistsException(e);
     }